requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.2.1
scikit-learn==1.3.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
numpy==1.26.4 